requirements.txt
# autopep8==2.0.4
# chardet==5.2.0
# flake8-coding==1.3.2
# frozendict==2.4.0
# func_timeout==4.3.5
# json-logic==0.6.3
# keyring==24.3.0
# looseversion==1.3.0
# matplotlib==3.9.0
openpyxl==3.1.2
pip==24.0
# pipreqs==0.4.13
plotly==5.19.0
py7zr==0.20.8
pydantic==2.5.3
pyfunctional==1.5.0
pylint==3.0.3
pymssql==2.2.11
PyMySQL==1.1.0
# pysvtools.crashdump_summarizer==0.0.70.600
pytest-benchmark==4.0.0
pytest-xdist==3.5.0
python-magic==0.4.27
rarfile==4.1
requests-ntlm==1.2.0
sh==2.0.6
SQLAlchemy==2.0.25
streamlit==1.35.0
# svtools.report==2.8.1.600
tqdm==4.66.4
tzlocal==5.2
validators==0.22.0
